Do math on integers, then convert to float in LineDrawingDecoration

+++ polet/src/main/java/com/fievx/polet/drawingDecoration/LineDrawingDecoration.kt
@@ -21,28 +21,28 @@ class LineDrawingDecoration (@ColorInt color: Int, val position: Position, val t
     override fun onDraw(rect: Rect, view: View, c: Canvas, parent: RecyclerView, state: RecyclerView.State) {
         when (position){
             Position.top ->
-                c.drawRect(view.left - rect.left.toFloat(),
-                    view.top - rect.top.toFloat(),
-                    view.right + rect.right.toFloat(),
-                    view.top - rect.top.toFloat() + thickness.toFloat(),
+                c.drawRect((view.left - rect.left).toFloat(),
+                    (view.top - rect.top).toFloat(),
+                    (view.right + rect.right).toFloat(),
+                    (view.top - rect.top + thickness.toFloat()).toFloat(),
                     paint)
             Position.bottom ->
-                c.drawRect(view.left - rect.left.toFloat(),
-                    view.bottom + rect.bottom.toFloat() - thickness.toFloat(),
-                    view.right + rect.right.toFloat(),
-                    view.bottom + rect.bottom.toFloat(),
+                c.drawRect((view.left - rect.left).toFloat(),
+                    (view.bottom + rect.bottom - thickness).toFloat(),
+                    (view.right + rect.right).toFloat(),
+                    (view.bottom + rect.bottom).toFloat(),
                     paint)
             Position.left ->
-                c.drawRect(view.left - rect.left.toFloat(),
-                    view.top - rect.top.toFloat(),
-                    view.left - rect.left.toFloat() + thickness.toFloat(),
-                    view.bottom + rect.bottom.toFloat(),
+                c.drawRect((view.left - rect.left).toFloat(),
+                    (view.top - rect.top).toFloat(),
+                    (view.left - rect.left + thickness).toFloat(),
+                    (view.bottom + rect.bottom).toFloat(),
                     paint)
             Position.right ->
-                c.drawRect(view.right + rect.right.toFloat() - thickness.toFloat(),
-                    view.top - rect.top.toFloat(),
-                    view.right + rect.right.toFloat(),
-                    view.bottom + rect.bottom.toFloat(),
+                c.drawRect((view.right + rect.right - thickness).toFloat(),
+                    (view.top - rect.top).toFloat(),
+                    (view.right + rect.right).toFloat(),
+                    (view.bottom + rect.bottom).toFloat(),
                     paint)
         }
     }
